Change black icon colors to white, add svg and ico versions
After contrasting the icon against other application icons, the black-on-purple did not look nearly as good, prompting the change to white-on-purple which looks better.

An SVG version has been added to enable changes to the icon in the long term, and the ico version has been added in preparation of the 2.5.* update on Windows.